backend/mysite/requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
psycopg2==2.9.5
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
